Refactor: move schemas into separate files
