+++ src/lib/openapi/index.ts
@@ -60,12 +60,10 @@ import { passwordSchema } from './spec/password-schema';
 import { patchesSchema } from './spec/patches-schema';
 import { patchSchema } from './spec/patch-schema';
 import { permissionSchema } from './spec/permission-schema';
-import {
-    playgroundFeatureSchema,
-    playgroundStrategySchema,
-    playgroundConstraintSchema,
-    playgroundSegmentSchema,
-} from './spec/playground-feature-schema';
+import { playgroundFeatureSchema } from './spec/playground-feature-schema';
+import { playgroundStrategySchema } from './spec/playground-strategy-schema';
+import { playgroundConstraintSchema } from './spec/playground-constraint-schema';
+import { playgroundSegmentSchema } from './spec/playground-segment-schema';
 import { playgroundRequestSchema } from './spec/playground-request-schema';
 import { playgroundResponseSchema } from './spec/playground-response-schema';
 import { projectEnvironmentSchema } from './spec/project-environment-schema';
