+++ src/lib/openapi/spec/playground-response-schema.ts
@@ -1,17 +1,14 @@
 import { FromSchema } from 'json-schema-to-ts';
 import { sdkContextSchema } from './sdk-context-schema';
 import { playgroundRequestSchema } from './playground-request-schema';
-import {
-    playgroundConstraintSchema,
-    playgroundFeatureSchema,
-    playgroundSegmentSchema,
-    playgroundStrategySchema,
-} from './playground-feature-schema';
+import { playgroundFeatureSchema } from './playground-feature-schema';
 import { constraintSchema } from './constraint-schema';
 import { parametersSchema } from './parameters-schema';
 import { variantSchema } from './variant-schema';
 import { overrideSchema } from './override-schema';
-
+import { playgroundConstraintSchema } from './playground-constraint-schema';
+import { playgroundSegmentSchema } from './playground-segment-schema';
+import { playgroundStrategySchema } from './playground-strategy-schema';
 
 export const playgroundResponseSchema = {
     $id: '#/components/schemas/playgroundResponseSchema',
@@ -44,7 +41,6 @@ export const playgroundResponseSchema = {
             sdkContextSchema,
             variantSchema,
             overrideSchema,
-
         },
     },
 } as const;
